steady version for test lab

+++ Chart.cpp
@@ -11,7 +11,7 @@
 #include <QMetaEnum>
 
 int MainWindow::encoderround1;    //catch the data from MainWindow.
-float MainWindow::motoround1;
+double MainWindow::motoround1;
 int MainWindow::biground1;
 
 Chart::Chart(QWidget *parent) :
